test(auth): Add `asyncTest` helper for integration tests
Various quirks in Flutter's `integration_test` package require that all asynchronous expectations be awaited. And integration tests run with `-d web-server` require that `testWidgets` be used instead of `test`.

This adds an `asyncTest` helper which can be used to work around these limitations and ensure that tests are run as expected.